Add files via upload
For this project, you’ll participate in a scenario. You work in the field of electoral politics, and you are specifically interested in ballot measures, where citizens vote directly to enact new laws. You are a GIS analyst for a nongovernmental organization that either supported or opposed (your choice) California’s 2012 ballot measure that would have required labeling of genetically modified organisms (GMOs) in food. Your task is to look at the broad voting patterns in California on this issue, knowing that the initiative did not gain enough votes to pass.